Adding some basic tests for timing and other such things.
